+++ tools/drawings/dr_016.py
@@ -1,16 +1,17 @@
-# -*- coding: utf-8 -*-
 """DR-028 街区景观品质分析图 — 多指标综合诊断
 展示 GVI（绿视率）、SVF（天空开敞度）、Enclosure（街道围合度）、Clutter（视觉杂乱度）
 四个核心街景品质指标的空间分布，量化评估人居环境景观品质痛点。
 """
 from pathlib import Path
+
 import geopandas as gpd
-import pandas as pd
-import numpy as np
+import matplotlib.colors as mcolors
 import matplotlib.font_manager as fm
 import matplotlib.patches as mpatches
 import matplotlib.patheffects as path_effects
-import matplotlib.colors as mcolors
+import numpy as np
+import pandas as pd
+
 from src.engines.spatial_engine import get_spatial_data
 
 ROOT = Path(__file__).resolve().parent.parent.parent
@@ -286,7 +287,7 @@ def draw_map(ax, roads, buildings, water, rails, key_plots, landuse, boundary,
     ax.text(x_end, y_bar + 1.5, "500m", color='#334155', ha='center', va='center',
             fontproperties=_font(font_prop, 10.0), zorder=4)
     scale_ratio = view_w / 0.31968
-    scale_rounded = int(round(scale_ratio / 500)) * 500
+    scale_rounded = round(scale_ratio / 500) * 500
     ax.text((x_start + x_end) / 2, y_bar - 1.6, f"比例尺 1:{scale_rounded}",
             color='#334155', ha='center', va='center',
             fontproperties=_font(font_prop, 10.5, 'bold'), zorder=4)
@@ -351,9 +352,10 @@ def draw_map(ax, roads, buildings, water, rails, key_plots, landuse, boundary,
 
         # Floating Windrose (Pure Black, 12.0 x 12.0) with soft white radial gradient backdrop
     try:
-        from PIL import Image as _PIL_Image
-        import numpy as _np
         from pathlib import Path as _Path
+
+        import numpy as _np
+        from PIL import Image as _PIL_Image
         _assets_dir = _Path(__file__).resolve().parent.parent.parent / "assets"
         _rose_path = _assets_dir / "长春市风玫瑰.png"
         if _rose_path.exists():
